Modulize a recyclerview for a list of PXLPhoto

+++ pixleesdk/src/main/java/com/pixlee/pixleesdk/util/NumberExt.kt
@@ -5,7 +5,7 @@ import android.content.res.Resources
 /**
  * Created by sungjun on 9/14/20.
  */
-val Int.dp: Int
-    get() = (this / Resources.getSystem().displayMetrics.density).toInt()
-val Int.px: Int
-    get() = (this * Resources.getSystem().displayMetrics.density).toInt()
+val Int.dp: Float
+    get() = (this / Resources.getSystem().displayMetrics.density)
+val Int.px: Float
+    get() = (this * Resources.getSystem().displayMetrics.density)
